requirements.txt
pytest==7.4.3
requests==2.31.0
pytest-xdist==3.3.1
allure-pytest==2.13.2
pydantic==1.10.13